requirements.txt
streamlit==1.32.0
pandas==2.1.4
openpyxl==3.1.2
numpy==1.26.4
sentence-transformers==2.7.0
faiss-cpu==1.12.0
pyyaml==6.0.1
unidecode==1.3.8
tqdm==4.66.1
huggingface-hub==0.34.4